move tests around, fix unit tests
