+++ paasta_itests/paasta_remote_run.feature
@@ -1,0 +1,6 @@
+Feature: Paasta remote-run
+  Scenario: We can run a job
+    Given a working paasta cluster, with docker registry docker.io
+      And a new adhoc config to be deployed
+     When we start an adhoc scheduler
+     Then it should eventually start 1 task
